myapi/requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.6
djangorestframework==3.14.0
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.98.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
lxml==4.9.1
magic-filter==1.0.12
multidict==6.0.4
oauth2client==4.1.3
Pillow==9.5.0
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.3.0
pydantic_core==2.6.3
PyDrive==1.3.1
pyparsing==3.1.1
pyTelegramBotAPI==4.13.0
python-pptx==0.6.22
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
XlsxWriter==3.1.3
yarl==1.9.2
gspread==5.11.3
